backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
openai
pillow
pytesseract
opencv-python
numpy
requests
